Capstone-project/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
distlib==0.3.8
Django==3.2.15
djangorestframework==3.14.0
djongo==1.3.6
filelock==3.13.1
pillow==10.2.0
pipenv==2023.11.15
platformdirs==4.1.0
pymongo==3.12.3
pytz==2023.3.post1
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.9.0
virtualenv==20.25.0
drf_yasg
dnspython==2.4.2
django-cors-headers==4.3.1
